backend/requirements.txt
uvicorn 
fastapi
openai
python-dotenv
sqlalchemy
asyncpg
bcrypt==4.0.1
alembic
psycopg2
passlib
fpdf2
markdown2
weasyprint